backend/requirements.txt
# 📦 HMIP-2026 Backend Dependencies
# File Location: backend/requirements.txt

# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
djongo==1.3.6
pymongo==4.6.1
sqlparse==0.4.4

# Authentication & Security
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
python-decouple==3.8
cryptography==42.0.0

# Async & WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1

# Celery for Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
kombu==5.3.4

# AI & Machine Learning
tensorflow==2.15.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.1
joblib==1.3.2

# Geospatial
geopy==2.4.1
haversine==2.8.0

# SMS & Notifications
twilio==8.11.0
python-dotenv==1.0.0

# API Documentation
drf-yasg==1.21.7
django-filter==23.5

# Image Processing
Pillow==10.2.0

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Data Validation
marshmallow==3.20.1

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.39.2

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.20.0
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Environment Management
python-dotenv==1.0.0

# Date & Time
python-dateutil==2.8.2
pytz==2024.1

# Utils
python-slugify==8.0.1
shortuuid==1.0.11